Add P2C tests
