+++ servicetalk-loadbalancer/src/test/java/io/servicetalk/loadbalancer/EagerNewRoundRobinLoadBalancerTest.java
@@ -15,7 +15,7 @@
  */
 package io.servicetalk.loadbalancer;
 
-class EagerNewRoundRobinLoadBalancerTest extends EagerRoundRobinLoadBalancerTest {
+class EagerNewRoundRobinLoadBalancerTest extends EagerLoadBalancerTest {
     @Override
     protected LoadBalancerBuilder<String, TestLoadBalancedConnection> baseLoadBalancerBuilder() {
         return LoadBalancers.<String, TestLoadBalancedConnection>builder(getClass().getSimpleName())
