+++ servicetalk-loadbalancer/src/test/java/io/servicetalk/loadbalancer/EagerP2CLoadBalancerTest.java
@@ -1,0 +1,9 @@
+package io.servicetalk.loadbalancer;
+
+public class EagerP2CLoadBalancerTest extends EagerLoadBalancerTest {
+
+    @Override
+    protected LoadBalancerBuilder<String, TestLoadBalancedConnection> baseLoadBalancerBuilder() {
+        return LoadBalancers.builder(getClass().getSimpleName());
+    }
+}
