+++ servicetalk-loadbalancer/src/test/java/io/servicetalk/loadbalancer/LingeringP2CLoadBalancerTest.java
@@ -1,0 +1,8 @@
+package io.servicetalk.loadbalancer;
+
+public class LingeringP2CLoadBalancerTest extends LingeringLoadBalancerTest {
+    @Override
+    protected LoadBalancerBuilder<String, TestLoadBalancedConnection> baseLoadBalancerBuilder() {
+        return LoadBalancers.builder(getClass().getSimpleName());
+    }
+}
